Project 0x11 C - printf

+++ p_int.c
@@ -6,10 +6,10 @@
  *
  * Return: int, byte count
  */
-int print_int(int num)
+int print_int(long int num)
 {
-	int x = num;
-	int i = 0;
+	long int x = num;
+	long int i = 0;
 
 	if (num)
 	{
